% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avs.r
\name{avs}
\alias{avs}
\title{Analisis Varians (ANOVA) dan Uji Lanjut (Post Hoc Test)}
\usage{
avs(excel_j, excel_k, sheet_n, sheet_k, anova_r, anova_p, anova_u, posthoc)
}
\arguments{
\item{excel_j}{Jalur ke file data excel.}

\item{excel_k}{Tipe kolom data excel.}

\item{sheet_n}{Nama sheet yang akan digunakan.}

\item{sheet_k}{Nama kolom hasil dalam sheet yang digunakan.}

\item{anova_r}{Tipe ANOVA ("RAK" atau "RAL").}

\item{anova_p}{Nama kolom perlakuan.}

\item{anova_u}{Nama kolom ulangan.}

\item{posthoc}{Tipe uji lanjut ("BNT", "BNJ", atau "DMRT").}
}
\value{
Hasil ANOVA dan Uji Lanjut.
}
\description{
Analisis Varians (ANOVA) dan Uji Lanjut (Post Hoc Test)
}
